+++ examples/local/xgboost_local.py
@@ -34,7 +34,7 @@ def run_experiment(on_travis: bool = False):
             print(configuration)
             for n_estimator in [8, 64]:
                 for subsample in [0.4, 1]:
-                    fidelity = {"n_estimators": n_estimator, "subsample": subsample}
+                    fidelity = {'n_estimators': n_estimator, 'subsample': subsample}
                     result_dict = b.objective_function(configuration.get_dictionary(),
                                                        fidelity=fidelity)
                     valid_loss = result_dict['function_value']
@@ -50,12 +50,12 @@ def run_experiment(on_travis: bool = False):
         print(f'Done, took totally {time()-start:.2f}')
 
 
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(prog='HPOlib CC Datasets', description='HPOlib3 on the CC18 data sets.',
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser(prog='HPOlib CC Datasets', description='HPOlib2 on the CC18 data sets.',
                                      usage='%(prog)s --array_id <task_id>')
 
     parser.add_argument('--on_travis', action='store_true',
-                        help='Flag to speed up the run on the continuous integration tool \"travis\". This flag can be'
+                        help='Flag to speed up the run on the continuous integration tool \'travis\'. This flag can be'
                              'ignored by the user')
 
     args = parser.parse_args()