+++ examples/w_optimizer/cartpole_bohb.py
@@ -117,7 +117,7 @@ def run_experiment(out_path, on_travis):
 if __name__ == '__main__':
     import argparse
     parser = argparse.ArgumentParser(prog='HPOlib - BOHB',
-                                     description='HPOlib3 with BOHB on Cartpole',
+                                     description='HPOlib2 with BOHB on Cartpole',
                                      usage='%(prog)s --out_path <string>')
     parser.add_argument('--out_path', default='./cartpole_smac_hb', type=str)
     parser.add_argument('--on_travis', action='store_true',