+++ changelog.md
@@ -1,3 +1,8 @@
+# 0.0.3
+  * improve forwarding exceptions in containerized benchmarks
+  * allow to set debug level with env variable
+  * rename everything to HPOlib2 
+
 # 0.0.2
   * add first set of benchmarks